+++ scripts/cli/diff-with-latest.ts
@@ -1,14 +1,8 @@
 import fs from "fs/promises";
 import { z } from "zod";
 import { createHash } from "node:crypto";
-import {
-  LOG_COLORS,
-  ScriptError,
-  ZBuildInfo,
-  ZContractInfo,
-  toAsyncResult,
-} from "../utils";
-import { retrieveFreshBuildInfo } from "./utils";
+import { ZBuildInfo, ZContractInfo, toAsyncResult } from "../utils";
+import { LOG_COLORS, retrieveFreshBuildInfo, ScriptError } from "./utils";
 
 /**
  * This script generates the differences between the artifacts generated by a fresh compilation and the ones of the `latest` release.
