+++ scripts/cli/utils.ts
@@ -49,3 +49,16 @@ export async function retrieveFreshBuildInfo() {
     content: contentResult.value,
   };
 }
+
+export const LOG_COLORS = {
+  log: "\x1b[0m%s\x1b[0m",
+  success: "\x1b[32m%s\x1b[0m",
+  error: "\x1b[31m%s\x1b[0m",
+  warn: "\x1b[33m%s\x1b[0m",
+};
+
+export class ScriptError extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}
